% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckME_Diallel.R
\name{CheckME_Diallel}
\alias{CheckME_Diallel}
\title{Checking the Properties of a ME-PDC}
\usage{
CheckME_Diallel(design)
}
\arguments{
\item{design}{Provide a ME-PDC}
}
\value{
Generates parameters of the designs along with C matrix, eigenvalues (EVs), degree of fractionations (DF) and canonical efficiency factor (CEF).
}
\description{
Checking the Properties of a ME-PDC
}
\examples{
library(MEDesigns)
design<-ME_PDC1(10)$ME_PDC
CheckME_Diallel(design)
}
