% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_CDC.R
\name{ME_CDC}
\alias{ME_CDC}
\title{ME-CDCs for Even Number of Lines}
\usage{
ME_CDC(lines)
}
\arguments{
\item{lines}{Number of Lines >=6}
}
\value{
ME-CDCs for an even number of lines along with their parameters, C matrices, eigenvalues (EVs) and canonical efficiency factor (CEF).
}
\description{
ME-CDCs for Even Number of Lines
}
\examples{
library(MEDesigns)
ME_CDC(6)
}
