% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_PDC1.R
\name{ME_PDC1}
\alias{ME_PDC1}
\title{ME-PDCs for Even Number of Lines}
\usage{
ME_PDC1(lines)
}
\arguments{
\item{lines}{Number of Lines >=6}
}
\value{
ME-PDCs for an even number of lines along with their parameters, C matrices, eigenvalues (EVs), degree of fractionations (DF) and canonical efficiency factor (CEF).
}
\description{
ME-PDCs for Even Number of Lines
}
\examples{
library(MEDesigns)
ME_PDC1(6)
}
