% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_PDC2.R
\name{ME_PDC2}
\alias{ME_PDC2}
\title{ME PDCs for Composite Number of Lines}
\usage{
ME_PDC2(p, q)
}
\arguments{
\item{p}{Any value (p>=3)}

\item{q}{Any value (q>=3)}
}
\value{
This function will provide ME-PDCs for a composite number, v(= pq) along with basic parameters, C matrix, eigenvalues (EVs), degree of fractionations (DF) and canonical efficiency factor (CEF).
}
\description{
ME PDCs for Composite Number of Lines
}
\examples{
library(MEDesigns)
ME_PDC2(3,3)
}
