\name{MEdata}
\alias{MEdata}
\docType{data}
\title{
Dataset for ME-PDC
}
\usage{data("MEdata")}
\description{
This is a sample dataset for user.
}
\format{
  A data frame with 40 observations on the following 5 variables.
  \describe{
    \item{\code{block}}{a numeric vector}
    \item{\code{line1}}{a numeric vector}
    \item{\code{line2}}{a numeric vector}
    \item{\code{cross_no}}{a numeric vector}
    \item{\code{yld}}{a numeric vector}
  }
}
\examples{
data(MEdata)
}
\keyword{datasets}
